Refactored Views and MainController
Made a new main controller which handles all interactions between controllers. This keeps things less coupled between controllers. Then I also refactored views so each window was its own view and easy to find. Then I also added a NurseController which handles functions which were being handled in multiple different places before unifying them into this class allows for easier debugging and additions.
