+++ src/views/duplicate_data_view.py
@@ -1,5 +1,5 @@
 import tkinter as tk
-from tkinter import ttk, messagebox
+from tkinter import ttk
 
 class DuplicateDataView:
     """
