+++ src/views/profile_view.py
@@ -78,9 +78,12 @@ class ProfileView:
         self.nurse_label.pack(anchor='w', pady=(5,10))
 
         # Buttons
-        tk.Button(frame, text="Assign Nurse", command=self.controller.assign_nurse).pack(pady=(10,5))
+        tk.Button(frame, text="Assign Nurse", command=self.assign_nurse).pack(pady=(10,5))
         tk.Button(frame, text="Copy Profile Info", command=self.controller.copy_to_clipboard).pack(pady=(5,5))
         tk.Button(frame, text="Export to PDF", command=self.controller.export_profile_to_pdf).pack(pady=(5,5))
+    
+    def assign_nurse(self):
+        self.controller.assign_nurse(self.child_data , self.update_nurse_info)
 
     def update_nurse_info(self, text):
         self.nurse_info_text = text
